refactor(size): split sleep-timer, chapter, and status-line modules; re-bless the baseline
- SleepTimerConfig moves to src/config/sleep_timer.rs so config.rs returns
  under its pinned cap.
- mpv chapter-list parsing moves to src/player/ipc/chapters.rs so incoming.rs
  returns under its pinned cap.
- The sleep-timer parity test moves to src/daemon/parity_tests/sleep_timer.rs.
- Seekbar chapter ticks move to src/ui/control_box/status_line.rs; the
  animated seekbar overlays stay in control_box.rs byte-identical.
- src/app/types.rs is pinned at 1528 lines in scripts/size-baseline.tsv (the
  tsv diff touches only that row); Msg/MouseTarget doc comments trimmed.
