+++ scripts/size-baseline.tsv
@@ -7,6 +7,7 @@
 1079	src/ui/control_box.rs
 1532	src/terminal_runtime/runner.rs
 1526	src/theme.rs
+1528	src/app/types.rs
 1516	src/runtime/tests.rs
 1516	src/transfer/engine/album_prefill.rs
 1515	src/download/tests.rs
